lspci: Replace unsigned long long type by u64 and %llx format by PCI_U64_FMT_X
pciutils already provides and uses u64 type together with PCI_U64_FMT_X
format macro for printing hex value of this type. So use u64 and
PCI_U64_FMT_X also on other few remaining places.

This change fixes printing hexadecimal 64-bit numbers by lspci on Window
systems independently of used compiler (MinGW or MSVC).

+++ ls-ecaps.c
@@ -68,12 +68,12 @@ cap_ltr(struct device *d, int where)
   snoop = get_conf_word(d, where + PCI_LTR_MAX_SNOOP);
   scale = cap_ltr_scale((snoop >> PCI_LTR_SCALE_SHIFT) & PCI_LTR_SCALE_MASK);
   printf("\t\tMax snoop latency: %lldns\n",
-	 ((unsigned long long)snoop & PCI_LTR_VALUE_MASK) * scale);
+	 ((u64)snoop & PCI_LTR_VALUE_MASK) * scale);
 
   nosnoop = get_conf_word(d, where + PCI_LTR_MAX_NOSNOOP);
   scale = cap_ltr_scale((nosnoop >> PCI_LTR_SCALE_SHIFT) & PCI_LTR_SCALE_MASK);
   printf("\t\tMax no snoop latency: %lldns\n",
-	 ((unsigned long long)nosnoop & PCI_LTR_VALUE_MASK) * scale);
+	 ((u64)nosnoop & PCI_LTR_VALUE_MASK) * scale);
 }
 
 static void
@@ -840,7 +840,7 @@ cap_l1pm(struct device *d, int where)
 	  if (scale > 5)
 	    printf(" LTR1.2_Threshold=<error>");
 	  else
-	    printf(" LTR1.2_Threshold=%lldns", BITS(val, 16, 10) * (unsigned long long) cap_ltr_scale(scale));
+	    printf(" LTR1.2_Threshold=%lldns", BITS(val, 16, 10) * (u64) cap_ltr_scale(scale));
 	}
       printf("\n");
     }
